Draft: change project structure
