+++ src/app/mod.rs
@@ -1,0 +1,57 @@
+pub mod connection;
+pub mod navigation;
+
+use connection::Connection;
+use core::fmt;
+use futures::TryFutureExt;
+use ratatui::{
+    style::{Style, Stylize},
+    text::Line,
+    widgets::{List, ListState},
+};
+use std::{cmp::min, mem, net::IpAddr, vec};
+use zookeeper_async::{Acl, Stat};
+
+use crate::{app_legacy::AppState, node_data::NodeData, tab::Tab};
+
+pub const BASE_RESOURCE: &str = "/";
+pub const CONFIRMATION_STRING: &str = "DELETE";
+#[derive(Default)]
+pub struct App {
+    pub state: AppState,
+    pub connection: Option<Connection>,
+    pub zk: Option<zookeeper_async::ZooKeeper>,
+    pub connection_input: String,
+    pub curr_tab: usize,
+    pub tabs: Vec<Tab>,
+
+    // -//- Moved to Tab struct
+    pub tab_data: Vec<String>,
+    pub list_state: ListState,
+    pub curr_resource: Option<String>, // selected node for current nest level
+    pub prev_resources: Vec<String>,   // prev resources: e.g. /zookeeper/config
+    pub current_node_stat: Option<Stat>,
+    pub message: String,
+    pub node_data: NodeData,
+    pub node_path_buf: String,
+    pub node_data_buf: String,
+    pub input_buf: String,
+}
+
+impl App {
+    pub fn new(connection: Connection) -> Self {
+        Self {
+            connection_input: connection.to_string(),
+            connection: Some(connection),
+            tabs: vec![Tab::default(), Tab::default(), Tab::default()],
+            curr_tab: 0usize,
+            ..Default::default()
+        }
+    }
+    pub fn connection_str(&self) -> String {
+        match self.connection {
+            Some(ref conn) => conn.to_string(),
+            None => "".to_owned(),
+        }
+    }
+}
