+++ src/main.rs
@@ -1,11 +1,13 @@
 pub mod app;
+pub mod app_legacy;
 pub mod cli;
 pub mod node_data;
 pub mod tab;
 pub mod ui;
 pub mod zk;
 
-use app::{App, AppState};
+use app::App;
+use app_legacy::AppState;
 use cli::parse_cli;
 use crossterm::event::{self, Event, KeyCode, KeyEventKind};
 use ratatui::{prelude::Backend, Terminal};
@@ -34,7 +36,7 @@ async fn run<B: Backend>(mut terminal: Terminal<B>, mut app: App) -> io::Result<
                 continue;
             }
             match app.state {
-                app::AppState::EstablishingConnection => match key.code {
+                app_legacy::AppState::EstablishingConnection => match key.code {
                     KeyCode::Esc => break Result::Ok(()),
                     KeyCode::Enter => {
                         let connection_string = app.connection_input.clone();
@@ -55,7 +57,7 @@ async fn run<B: Backend>(mut terminal: Terminal<B>, mut app: App) -> io::Result<
                     KeyCode::Char('q') => break Result::Ok(()),
                     _ => {}
                 },
-                app::AppState::EditingConnection if key.kind == KeyEventKind::Press => {
+                app_legacy::AppState::EditingConnection if key.kind == KeyEventKind::Press => {
                     match key.code {
                         KeyCode::Esc => app.state = AppState::EstablishingConnection,
                         KeyCode::Enter => {
@@ -70,7 +72,7 @@ async fn run<B: Backend>(mut terminal: Terminal<B>, mut app: App) -> io::Result<
                         _ => {}
                     }
                 }
-                app::AppState::Tab => match key.code {
+                app_legacy::AppState::Tab => match key.code {
                     KeyCode::Char('j') | KeyCode::Down => {
                         app.next();
                         app.curr_resource = app.selected_resource();
